Update MyGradeBook.java
Made a few fixes I found with BlackBox Tests:

Removed uses of KeySet when simply referencing the list of assignments will do (because every student has a grade for every assignment). I think I got all of them, but if you see one somewhere, see if it can be replaced.

Changed a use of == when it should have been equals().

One of the output methods wasn't accumulating properly - at a few points in the method, instead of output = output + X, it was output = X, removing everything prior to that.